backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Exchange APIs
ccxt==4.2.25

# AI/ML
langchain==0.1.4
google-generativeai==0.3.2
transformers==4.37.2
torch==2.2.0
vaderSentiment==3.3.2

# Data processing
pandas==2.2.0
numpy==1.26.3

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Environment and config
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
